Move Cached::Reference::BelongsTo to Cached::BelongsTo
