+++ lib/identity_cache.rb
@@ -9,11 +9,11 @@ require "identity_cache/cached"
 require "identity_cache/cached/prefetcher"
 require "identity_cache/cached/embedded_fetching"
 require "identity_cache/cached/association"
+require "identity_cache/cached/belongs_to"
 require "identity_cache/cached/recursive/association"
 require "identity_cache/cached/recursive/has_one"
 require "identity_cache/cached/recursive/has_many"
 require "identity_cache/cached/reference/association"
-require "identity_cache/cached/reference/belongs_to"
 require "identity_cache/cached/reference/has_one"
 require "identity_cache/cached/reference/has_many"
 require "identity_cache/expiry_hook"
