+++ lib/identity_cache/belongs_to_caching.rb
@@ -23,10 +23,7 @@ module IdentityCache
           )
         end
 
-        cached_belongs_to = Cached::Reference::BelongsTo.new(
-          association,
-          reflection: reflection,
-        )
+        cached_belongs_to = Cached::BelongsTo.new(association, reflection: reflection)
 
         self.cached_belongs_tos[association] = cached_belongs_to.tap(&:build)
       end
