+++ lib/identity_cache/cached/recursive/association.rb
@@ -40,6 +40,7 @@ module IdentityCache
 
         def fetch(records)
           fetch_embedded(records)
+
           records.flat_map(&cached_accessor_name.to_sym).tap(&:compact!)
         end
